+++ CHANGELOG.md
@@ -20,6 +20,7 @@ $ cd gvm-tools && git log
 - Added --ssh-password switch for ssh connection [PR 140](https://github.com/greenbone/gvm-tools/pull/140)
 - Added a new console line interface `gvm-script` for only running GMP and OSP
   scripts without opening a python shell [PR 152](https://github.com/greenbone/gvm-tools/pull/152)
+- Forbid to run any gvm-tools cli as root user [PR 183](https://github.com/greenbone/gvm-tools/pull/183)
 
 ### Changed
 - Improved error messages if unix socket could not be found [PR 78](https://github.com/greenbone/python-gvm/pull/78)