+++ CHANGELOG.md
@@ -32,6 +32,8 @@ $ cd gvm-tools && git log
 - Renamed --ssh-user switch to --ssh-username [PR 140](https://github.com/greenbone/gvm-tools/pull/140)
 - Update `gvmtools.get_version` to return a fully compliant [PEP 440](https://www.python.org/dev/peps/pep-0440/)
   version string [PR 150](https://github.com/greenbone/gvm-tools/pull/150)
+- Refresh the dependencies specified via the `Pipfile.lock` file to their latest
+  versions [PR 186](https://github.com/greenbone/gvm-tools/pull/186)
 
 ### Deprecated
 - Only running scripts with gvm-pyshell is deprecated [PR 152](https://github.com/greenbone/gvm-tools/pull/152)