+++ CHANGELOG.md
@@ -21,6 +21,7 @@ $ cd gvm-tools && git log
 - Added a new console line interface `gvm-script` for only running GMP and OSP
   scripts without opening a python shell [PR 152](https://github.com/greenbone/gvm-tools/pull/152)
 - Forbid to run any gvm-tools cli as root user [PR 183](https://github.com/greenbone/gvm-tools/pull/183)
+- Added error message if invalid XML is passed to `gvm-cli` [PR 198](https://github.com/greenbone/gvm-tools/pull/198)
 
 ### Changed
 - Improved error messages if unix socket could not be found [PR 78](https://github.com/greenbone/python-gvm/pull/78)